+++ app/jobs/scheduled_job_runner.rb
@@ -47,7 +47,7 @@ class ScheduledJobRunner < ScheduledJob
       ProcessGraduatedNotificationJob,
       ProcessHotSheetJob,
       ScheduledAutocompleteCheckJob,
-      ScheduledBikePossiblyFoundNotificationJob,
+      BikeJobs::ScheduledBikePossiblyFoundNotificationJob,
       ScheduledSearchForExternalRegistryBikesJob,
       ScheduledStoreLogSearchesJob,
       Spreadsheets::TsvCreatorJob,
